+++ src/main/java/wooteco/subway/admin/controller/PathController.java
@@ -4,9 +4,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.admin.dto.PathRequest;
-import wooteco.subway.admin.dto.PathRequestWithId;
-import wooteco.subway.admin.dto.PathResponse;
+import wooteco.subway.admin.dto.path.PathRequest;
+import wooteco.subway.admin.dto.path.PathRequestWithId;
+import wooteco.subway.admin.dto.path.PathResponse;
 import wooteco.subway.admin.service.PathService;
 import wooteco.subway.admin.service.StationService;
