+++ src/main/java/wooteco/subway/admin/dto/path/PathRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.path;
 
 public class PathRequest {
     private String sourceName;
