+++ src/main/java/wooteco/subway/admin/dto/path/PathRequestWithId.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.path;
 
 public class PathRequestWithId {
     private Long sourceId;
