+++ src/main/java/wooteco/subway/admin/dto/station/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.station;
 
 import wooteco.subway.admin.domain.Station;
