+++ src/main/java/wooteco/subway/admin/exception/NotFoundStationException.java
@@ -2,9 +2,9 @@ package wooteco.subway.admin.exception;
 
 public class NotFoundStationException extends IllegalArgumentException {
 
-    public static final String ERROR_MESSAGE = "존재하지 않는 역입니다.";
+    public static final String ERROR_MESSAGE = "존재하지 않는 역입니다 [%s]";
 
-    public NotFoundStationException() {
-        super(ERROR_MESSAGE);
+    public NotFoundStationException(String name) {
+        super(String.format(ERROR_MESSAGE, name));
     }
 }
