+++ src/main/java/wooteco/subway/admin/service/GraphService.java
@@ -1,17 +1,15 @@
 package wooteco.subway.admin.service;
 
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Service;
-
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.dto.PathType;
+import wooteco.subway.admin.dto.path.PathType;
 import wooteco.subway.admin.exception.NotFoundPathException;
+
+import java.util.List;
+import java.util.Objects;
 
 @Service
 public class GraphService {
@@ -20,7 +18,7 @@ public class GraphService {
             = new WeightedMultigraph(DefaultWeightedEdge.class);
         lines.stream()
             .flatMap(it -> it.getLineStationsId().stream())
-            .forEach(it -> graph.addVertex(it));
+            .forEach(graph::addVertex);
         lines.stream()
             .flatMap(it -> it.getStations().stream())
             .filter(it -> Objects.nonNull(it.getPreStationId()))
