+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -2,11 +2,11 @@ package wooteco.subway.admin.acceptance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.admin.dto.StationResponse;
+import wooteco.subway.admin.dto.station.StationResponse;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 public class StationAcceptanceTest extends AcceptanceTest {
     @DisplayName("지하철역을 관리한다")
